Correct degree character in the Fahrenheit unit
Currently the unit is set to `˚F` which uses ascii character 730 for the degree symbol. Home Assistant uses `°F` with ascii character 176 for the degree symbol.

When the symbols are different we can't combine different measurements into the same graph in Home Assistant.

I've looked into changing the unit of measurement directly in Home Assistant but there is no option showing up for that, and I believe that it's because this difference in unit is affecting that as well.